+++ packages/web-vue/components/button/button.vue
@@ -95,9 +95,9 @@ export default defineComponent({
     status: {
       type: String as PropType<Status>,
       default: 'normal',
-      validator: (value: any) => {
-        return STATUSES.includes(value);
-      },
+      // validator: (value: any) => {
+      //   return STATUSES.includes(value);
+      // },
     },
     /**
      * @zh 按钮的尺寸
@@ -149,7 +149,9 @@ export default defineComponent({
      */
     href: String,
     // for JSX
-    onClick: [Function, Array] as PropType<EmitType<(ev: MouseEvent) => void>>,
+    onClick: {
+      type: [Function, Array] as PropType<EmitType<(ev: MouseEvent) => void>>,
+    },
   },
   emits: [
     /**