+++ packages/web-vue/components/table/table.tsx
@@ -5,7 +5,15 @@ import type {
   Slot,
   VNode,
 } from 'vue';
-import { computed, defineComponent, onMounted, ref, toRefs, watch } from 'vue';
+import {
+  computed,
+  defineComponent,
+  onMounted,
+  reactive,
+  ref,
+  toRefs,
+  watch,
+} from 'vue';
 import { getPrefixCls } from '../_utils/global-config';
 import { off, on } from '../_utils/dom';
 import type { Size } from '../_utils/constant';
@@ -199,11 +207,28 @@ export default defineComponent({
     virtualListProps: {
       type: Object as PropType<VirtualListProps>,
     },
+    /**
+     * @zh 单元格合并方法（索引从数据项开始计数）
+     * @en Cell merge method (The index starts counting from the data item)
+     * @version 2.10.0
+     */
+    spanMethod: {
+      type: Function as PropType<
+        (data: {
+          record: TableData;
+          column: TableColumn;
+          rowIndex: number;
+          columnIndex: number;
+        }) => { rowspan?: number; colspan?: number } | void
+      >,
+    },
     components: {
       type: Object as PropType<TableComponents>,
     },
     // for JSX
-    onExpand: [Function, Array] as PropType<EmitType<(rowKey: string) => void>>,
+    onExpand: {
+      type: [Function, Array] as PropType<EmitType<(rowKey: string) => void>>,
+    },
     onExpandedChange: {
       type: [Function, Array] as PropType<
         EmitType<(rowKeys: string[]) => void>
@@ -250,61 +275,76 @@ export default defineComponent({
     /**
      * @zh 点击展开行时触发
      * @en Triggered when a row is clicked to expand
+     * @param {string} rowKey
      */
     'expand',
     /**
      * @zh 已展开的数据行发生改变时触发
      * @en Triggered when the expanded data row changes
+     * @param {string[]} rowKeys
      */
     'expandedChange',
     /**
      * @zh 点击行选择器时触发
      * @en Triggered when the row selector is clicked
+     * @param {string[]} rowKeys
      */
     'select',
     /**
      * @zh 点击全选选择器时触发
      * @en Triggered when the select all selector is clicked
+     * @param {boolean} checked
      */
     'selectAll',
     /**
      * @zh 已选择的数据行发生改变时触发
      * @en Triggered when the selected data row changes
+     * @param {string[]} rowKeys
      */
     'selectionChange',
     /**
      * @zh 排序规则发生改变时触发
      * @en Triggered when the collation changes
+     * @param {string} dataIndex
+     * @param {string} direction
      */
     'sorterChange',
     /**
      * @zh 过滤选项发生改变时触发
      * @en Triggered when the filter options are changed
+     * @param {string} dataIndex
+     * @param {string[]} filteredValues
      */
     'filterChange',
     /**
      * @zh 表格分页发生改变时触发
      * @en Triggered when the table pagination changes
+     * @param {number} page
      */
     'pageChange',
     /**
      * @zh 表格每页数据数量发生改变时触发
      * @en Triggered when the number of data per page of the table changes
+     * @param {number} pageSize
      */
     'pageSizeChange',
     /**
      * @zh 点击单元格时触发
      * @en Triggered when a cell is clicked
+     * @param {TableData} record
+     * @param {TableColumn} column
      */
     'cellClick',
     /**
      * @zh 点击行数据时触发
      * @en Triggered when row data is clicked
+     * @param {TableData} record
      */
     'rowClick',
     /**
      * @zh 点击表头数据时触发
      * @en Triggered when the header data is clicked
+     * @param {TableColumn} column
      */
     'headerClick',
   ],
@@ -318,6 +358,11 @@ export default defineComponent({
    * @en Expand row content
    * @slot expand-row
    * @binding {TableData} record
+   */
+  /**
+   * @zh 表格底部
+   * @en Table Footer
+   * @slot footer
    */
   setup(props, { emit, slots }) {
     const { rowKey } = toRefs(props);
@@ -1029,6 +1074,7 @@ export default defineComponent({
 
       return (
         <button
+          type="button"
           class={`${prefixCls}-expand-btn`}
           onClick={() => handleExpand(currentKey)}
         >
@@ -1040,6 +1086,45 @@ export default defineComponent({
         </button>
       );
     };
+
+    // [row, column]
+    const tableSpan = computed(() => {
+      const data: Record<string, [number, number]> = {};
+      flattenData.value.forEach((record, rowIndex) => {
+        dataColumns.value.forEach((column, columnIndex) => {
+          const { rowspan = 1, colspan = 1 } =
+            props.spanMethod?.({
+              record,
+              column,
+              rowIndex,
+              columnIndex,
+            }) ?? {};
+          if (rowspan > 1 || colspan > 1) {
+            data[`${rowIndex}-${columnIndex}`] = [rowspan, colspan];
+          }
+        });
+      });
+
+      return data;
+    });
+
+    const removedCells = computed(() => {
+      const data: string[] = [];
+      for (const indexKey of Object.keys(tableSpan.value)) {
+        const indexArray = indexKey.split('-').map((item) => Number(item));
+        const span = tableSpan.value[indexKey];
+        for (let i = 1; i < span[0]; i++) {
+          data.push(`${indexArray[0] + i}-${indexArray[1]}`);
+          for (let j = 1; j < span[1]; j++) {
+            data.push(`${indexArray[0] + i}-${indexArray[1] + j}`);
+          }
+        }
+        for (let i = 1; i < span[1]; i++) {
+          data.push(`${indexArray[0]}-${indexArray[1] + i}`);
+        }
+      }
+      return data;
+    });
 
     const renderRecord = (
       record: TableData,
@@ -1110,6 +1195,15 @@ export default defineComponent({
                     }
                   : undefined;
 
+              const cellId = `${rowIndex}-${index}`;
+              const [rowspan, colspan] = tableSpan.value[
+                `${rowIndex}-${index}`
+              ] ?? [1, 1];
+
+              if (removedCells.value.includes(cellId)) {
+                return null;
+              }
+
               return (
                 <Td
                   key={`td-${index}`}
@@ -1123,6 +1217,8 @@ export default defineComponent({
                   column={column}
                   operations={operations.value}
                   dataColumns={dataColumns.value}
+                  rowSpan={rowspan}
+                  colSpan={colspan}
                   {...extraProps}
                   onClick={(e: Event) => handleCellClick(record, column)}
                 >