+++ packages/web-vue/components/textarea/textarea.vue
@@ -1,5 +1,9 @@
 <template>
-  <div :class="wrapperCls" @mousedown="handleMousedown">
+  <div
+    v-bind="getWrapperAttrs($attrs)"
+    :class="wrapperCls"
+    @mousedown="handleMousedown"
+  >
     <div
       v-if="autoSize"
       ref="mirrorRef"
@@ -11,9 +15,9 @@
     <resize-observer @resize="handleResize">
       <textarea
         ref="textareaRef"
-        v-bind="$attrs"
+        v-bind="getTextareaAttrs($attrs)"
         :disabled="disabled"
-        :class="cls"
+        :class="prefixCls"
         :style="textareaStyle"
         :value="computedValue"
         :placeholder="placeholder"
@@ -25,6 +29,7 @@
         @compositionend="handleComposition"
       />
     </resize-observer>
+    <slot name="suffix" />
     <div v-if="maxLength && showWordLimit" :class="`${prefixCls}-word-limit`">
       {{ getTextLength(computedValue) }}/{{ maxLength }}
     </div>
@@ -56,8 +61,11 @@ import IconHover from '../_components/icon-hover.vue';
 import IconClose from '../icon/icon-close';
 import { getPrefixCls } from '../_utils/global-config';
 import { getSizeStyles } from './utils';
-import { isFunction } from '../_utils/is';
+import { isFunction, isObject } from '../_utils/is';
 import { EmitType } from '../_utils/types';
+import { omit } from '../_utils/omit';
+import { INPUT_EVENTS } from '../_utils/constant';
+import pick from '../_utils/pick';
 
 export default defineComponent({
   name: 'Textarea',
@@ -125,7 +133,9 @@ export default defineComponent({
      * @en Whether to make the textarea adapt to the height of the content
      */
     autoSize: {
-      type: Boolean,
+      type: [Boolean, Object] as PropType<
+        boolean | { minRows?: number; maxRows?: number }
+      >,
       default: false,
     },
     /**
@@ -136,15 +146,25 @@ export default defineComponent({
       type: Function as PropType<(value: string) => number>,
     },
     // for JSX
-    onInput: [Function, Array] as PropType<
-      EmitType<(value: string, ev: Event) => void>
-    >,
-    onChange: [Function, Array] as PropType<
-      EmitType<(value: string, ev: Event) => void>
-    >,
-    onClear: [Function, Array] as PropType<EmitType<(ev: MouseEvent) => void>>,
-    onFocus: [Function, Array] as PropType<EmitType<(ev: FocusEvent) => void>>,
-    onBlur: [Function, Array] as PropType<EmitType<(ev: FocusEvent) => void>>,
+    onInput: {
+      type: [Function, Array] as PropType<
+        EmitType<(value: string, ev: Event) => void>
+      >,
+    },
+    onChange: {
+      type: [Function, Array] as PropType<
+        EmitType<(value: string, ev: Event) => void>
+      >,
+    },
+    onClear: {
+      type: [Function, Array] as PropType<EmitType<(ev: MouseEvent) => void>>,
+    },
+    onFocus: {
+      type: [Function, Array] as PropType<EmitType<(ev: FocusEvent) => void>>,
+    },
+    onBlur: {
+      type: [Function, Array] as PropType<EmitType<(ev: FocusEvent) => void>>,
+    },
   },
   emits: [
     'update:modelValue',
@@ -174,7 +194,7 @@ export default defineComponent({
      */
     'blur',
   ],
-  setup(props, { emit }) {
+  setup(props, { emit, attrs }) {
     const prefixCls = getPrefixCls('textarea');
 
     const textareaRef = ref<HTMLInputElement>();
@@ -292,33 +312,67 @@ export default defineComponent({
       }
     );
 
+    const getWrapperAttrs = (attr: Record<string, any>) =>
+      omit(attrs, INPUT_EVENTS);
+    const getTextareaAttrs = (attr: Record<string, any>) =>
+      pick(attrs, INPUT_EVENTS);
+
     const wrapperCls = computed(() => [
       `${prefixCls}-wrapper`,
-      {
-        [`${prefixCls}-scroll`]: isScroll.value,
-      },
-    ]);
-
-    const cls = computed(() => [
-      prefixCls,
       {
         [`${prefixCls}-focus`]: focused.value,
         [`${prefixCls}-disabled`]: props.disabled,
         [`${prefixCls}-error`]: props.error,
+        [`${prefixCls}-scroll`]: isScroll.value,
       },
     ]);
 
     let styleDeclaration: CSSStyleDeclaration;
 
+    const lineHeight = ref<number>(0);
+    const outerHeight = ref<number>(0);
+    const minHeight = computed(() => {
+      if (!isObject(props.autoSize) || !props.autoSize.minRows) {
+        return 0;
+      }
+      return props.autoSize.minRows * lineHeight.value + outerHeight.value;
+    });
+    const maxHeight = computed(() => {
+      if (!isObject(props.autoSize) || !props.autoSize.maxRows) {
+        return 0;
+      }
+      return props.autoSize.maxRows * lineHeight.value + outerHeight.value;
+    });
+
     const getMirrorStyle = () => {
-      mirrorStyle.value = getSizeStyles(styleDeclaration);
+      const styles = getSizeStyles(styleDeclaration);
+
+      lineHeight.value = Number.parseInt(styles['line-height'], 10);
+      outerHeight.value =
+        Number.parseInt(styles['border-width'], 10) * 2 +
+        Number.parseInt(styles['padding-top'], 10) +
+        Number.parseInt(styles['padding-bottom'], 10);
+
+      mirrorStyle.value = styles;
+
       nextTick(() => {
         const mirrorHeight = mirrorRef.value?.offsetHeight;
 
+        let height = mirrorHeight ?? 0;
+        let overflow = 'hidden';
+
+        if (minHeight.value && height < minHeight.value) {
+          height = minHeight.value;
+        }
+        if (maxHeight.value && height > maxHeight.value) {
+          height = maxHeight.value;
+          overflow = 'auto';
+        }
+
         textareaStyle.value = {
-          height: `${mirrorHeight}px`,
+          height: `${height}px`,
           resize: 'none',
-          overflow: 'hidden',
+          overflow,
         };
       });
     };
@@ -367,7 +421,6 @@ export default defineComponent({
     return {
       prefixCls,
       wrapperCls,
-      cls,
       textareaRef,
       textareaStyle,
       mirrorRef,
@@ -375,6 +428,8 @@ export default defineComponent({
       computedValue,
       showClearBtn,
       getTextLength,
+      getWrapperAttrs,
+      getTextareaAttrs,
       handleInput,
       handleFocus,
       handleBlur,