+++ packages/web-vue/components/tree-select/tree-select.vue
@@ -5,11 +5,13 @@
     trigger="click"
     position="bl"
     :popup-offset="4"
+    animation-name="slide-dynamic-origin"
     :prevent-focus="true"
     v-bind="triggerProps"
     :disabled="disabled"
     :popup-visible="panelVisible"
     :popup-container="popupContainer"
+    auto-fit-transform-origin
     @popupVisibleChange="onVisibleChange"
   >
     <slot name="trigger">
@@ -23,6 +25,7 @@
         :size="size"
         :max-tags="maxTags"
         :disabled="disabled"
+        :opened="panelVisible"
         :error="error"
         :border="border"
         :placeholder="placeholder"
@@ -336,26 +339,32 @@ export default defineComponent({
       >,
     },
     // for JSX
-    onChange: [Function, Array] as PropType<
-      EmitType<
-        (
-          selectedValue:
-            | string
-            | number
-            | LabelValue
-            | Array<string | number>
-            | LabelValue[]
-            | undefined
-        ) => void
-      >
-    >,
-    onPopupVisibleChange: [Function, Array] as PropType<
-      EmitType<(popupVisible: boolean) => void>
-    >,
-    onSearch: [Function, Array] as PropType<
-      EmitType<(searchValue: string) => void>
-    >,
-    onClear: [Function, Array] as PropType<EmitType<() => void>>,
+    onChange: {
+      type: [Function, Array] as PropType<
+        EmitType<
+          (
+            selectedValue:
+              | string
+              | number
+              | LabelValue
+              | Array<string | number>
+              | LabelValue[]
+              | undefined
+          ) => void
+        >
+      >,
+    },
+    onPopupVisibleChange: {
+      type: [Function, Array] as PropType<
+        EmitType<(popupVisible: boolean) => void>
+      >,
+    },
+    onSearch: {
+      type: [Function, Array] as PropType<
+        EmitType<(searchValue: string) => void>
+      >,
+    },
+    onClear: { type: [Function, Array] as PropType<EmitType<() => void>> },
   },
   emits: [
     /**