+++ packages/web-vue/components/cascader/CHANGELOG.md
@@ -2,6 +2,32 @@
 changelog: true
 ```
 
+## 2.10.0
+
+`2021-12-10`
+
+### 💎 Optimization
+
+- When the input box is editable, clicking will not close the drop-down menu ([#348](https://github.com/arco-design/arco-design-vue/pull/348))
+
+### 🆕 Feature
+
+- Add checkStrictly prop ([#349](https://github.com/arco-design/arco-design-vue/pull/349))
+
+
+## 2.8.0
+
+`2021-12-01`
+
+### 🆕 Feature
+
+- Add support for `tagProps` ([#307](https://github.com/arco-design/arco-design-vue/pull/307))
+
+### 💅 Style
+
+- Remove the option to select the bold effect ([#308](https://github.com/arco-design/arco-design-vue/pull/308))
+
+
 ## 2.6.0
 
 `2021-11-19`